fix(monitor_stack): Remove circular dependency
Due to all logic being in __init__.py, when importing ui submodule,
 which doesnt have any dependencies,
 you could get ciruclar dependency as everything from init would be loaded.
Fix by moving all __init__.py into service.py module
